refactor(waterfall): move deal-by-deal ledger to shared for server reuse
Relocate calculateAmericanWaterfallLedger and its types from
client/src/lib/waterfall/american-ledger.ts to
shared/lib/waterfall/american-ledger.ts (byte-identical move), leaving a
re-export shim at the old client path (precedent:
client/src/core/cohorts/CohortEngine.ts). Update the two truth-case
import sites to the shared path.

Closes the @/ alias loophole in the ESLint client-boundary rule (issue
#1176 waterfall-gate finding W6); runner.test.ts direct-invocation site
per finding W11. PLAN_61 Wave F Task 16.1.

Verification: npm run check, npm run lint, npm run phoenix:truth -- all
14 waterfall-ledger truth cases (L01-L14) green with zero numeric
change; analytics-waterfall suite green through the shim.
